chore: extract useQueryWidgetSidePanel
